+++ src/main/java/com/finalpre/quickshare/service/EmailService.java
@@ -1,5 +1,5 @@
 package com.finalpre.quickshare.service;
 
 public interface EmailService {
-    void sendVerificationCode(String email, String code);
+    void sendVerificationCode(String email, String code, String locale);
 }