+++ src/main/java/com/finalpre/quickshare/service/VerificationCodeService.java
@@ -1,6 +1,6 @@
 package com.finalpre.quickshare.service;
 
 public interface VerificationCodeService {
-    String generateAndSendCode(String email, String recaptchaToken);
+    String generateAndSendCode(String email, String recaptchaToken, String locale);
     boolean verifyCode(String email, String code);
 }